add cc license
this remains under the condition that the code found within this repository is not part of any source code from another project that also requires another license. the conditions found in this file allow for the code to be both used and adapted but not claimed.

NOTE: code found here is owned by me, and me alone (unless otherwise stated with references or attributes found in the folder's readme or file).